Reorganize persistence folder
