+++ pylon-test/src/main/java/io/github/pylonmc/pylon/test/test/block/BlockStorageSimpleTest.java
@@ -3,7 +3,7 @@ package io.github.pylonmc.pylon.test.test.block;
 import io.github.pylonmc.pylon.core.block.PylonBlock;
 import io.github.pylonmc.pylon.core.block.PylonBlockSchema;
 import io.github.pylonmc.pylon.core.persistence.BlockStorage;
-import io.github.pylonmc.pylon.core.persistence.PylonDataReader;
+import io.github.pylonmc.pylon.core.persistence.pdc.PylonDataReader;
 import io.github.pylonmc.pylon.core.test.GameTestConfig;
 import io.github.pylonmc.pylon.test.PylonTest;
 import io.github.pylonmc.pylon.test.base.GameTest;
